Improved the "GetAttribute" method
In native .NET, the "Attribute" method supports specifying a string with the name of the attribute since it is automatically casted to "XName" using operator overloading. With the "GetAttribute" extension methods in this project, there is overload taking a string parameter for an XPath expression, so the automatic casting to "XName" doesn't occur. Therefore, if the expression looks like the name of an attribute, we manually cast it to an instance of "XName".