Changing the Newtonsoft.Json dependency a bit
Umbraco has a very version specific dependency for Newtonsoft.Json, so we need to support the lowest common denominator - at least from Umbraco 7.2 and up. This means that the .NET framework versions will have a dependency for Newtonsoft.Json 6.0.8, while the .NET Standard versions will have a dependency for Newtonsoft.Json 9.0.1 (as 9.x is the first version supporting .NET Standard).