Improved support for code inspection with JetBrains.Annotations
This will help ReSharper give better suggestions when using the methods and parameters that now has been decorated with attributes from the JetBrains.Annotations package.

JetBrains.Annotations is installed as a dev assembly, meaning that it will not be installed along Skybrud.Essentials, but you still get the benefits of you have ReSharper installed. If use reflection against the decorated methods and parameters, you'll however have to install the JetBrains.Annotations package manually in your project to avoid exceptions about the used attributes not being found.