Introduced new StringJsonConverter and NameValueCollectionJsonConverter classes
StringJsonConverter replaces the ToStringJsonConverter class, and serializes any .NET type to it's corresponding string representation. For NameValueCollection that means an URL encoded string, but for any other types, the value will be the same as calling the ToString method with an invariant culture. Generally the converter doesn't support deserialization, as must types can't be converted from a string to an object. But specically, the converter currently supports deserializing to NameValueCollection and any enum type.

The NameValueCollectionJsonConverter class serializes a NameValueCollection to a corresponding JObject where each key value pair in the value collection is added as properties and in the resulting JObject. The converter also supports deserializing a JObject back to a corresponding NameValueCollection.