Introduced new "StringExtensions.IfNullOrWhiteSpace" extension methods
For the first method, if it's passed a value that is either null, empty or white space, the specified "fallback" value is returned instead.

The second method works in a similar, but supports a "fallback" function instead. This is useful if the fallback value requires expensive lookups, as the function is only invoked when needed.