Added "IsBoolean" method overloads to the "StringExtensions" class
The first method overload takes a string as it's only parameter, and returns whether the string value can be parsed and converted into a bool equivalent.

The second method overload does the same, but has an addtional parameter that uses the "out" keyword to expose the "bool?" value that resulted from the conversion.

See #30