Updated NuGet dependencies
For the lifetime of this package, .NET Framework versions of the package has had a dependency for "Newtonsoft.Json" 6.0.8 under the basis that the package supported old versions not forcing people to upgrade Newtonsoft.Json, but also allowed people to use newer versions as they would probably already have Newtonsoft.Json installed. A lot of time has passed since then, so now might the time to bump the lower requirement to 9.0.1.

To provide code analysis for older frameworks, recent versions of this package have also had a dependency for "Nullable", but not for all target frameworks. Since the dependency is private, it shouldn't hurt having the dependency for all target frameworks. This also makes the .csproj file a bit more clean.