Introduced new "ReadOnlySet" class and "ReadOnlySet.Empty<T>" static method
Similar to the "Enumerable.Empty<T>" and "Array.Empty" method in .NET, Skybrud.Essentials now offers a method to get an empty immutable instance of "IReadOnlySet". As the "IReadOnlySet" interface is only supported from .NET 5 and up, the new class and method is also only available for .NET 5 and up, but not .NET Framework and .NET Standard.