Updated the "EnumUtils.TryParseEnum" methods to not throw exceptions
The general idea with "TryParseX" methods is that they should fail silently, and then return a boolean value to indicate whether the input value could be parsed. For the "TryParseEnum", they would in some cases throw an exception - e.g. if the input string was null. With this commit, they will return false instead and thus fail silently.

See #54