Better exception handling for "EnumUtils.Parse" method
In previous releases, the method would throw an "ArgumentNullException" regardless of the input string being null or just empty/white space. As of this commit, the method will instead throw the lesser "ArgumentException". This matches .NET's behavior, and it also ensures that the exception message matches the actual issue.

Ideally we should just use .NET's "ArgumentException.ThrowIfNullOrWhiteSpace" method, but this is first availble from .NET 8, which we currently doesn't target directly.