Fixed misc generic method constraints when working with enums
IIRC older versions of .NET Framework (and C#) didn't allow using "Enum" as a generic constraint, so one had to use "struct" instead. When support for "Enum" was lated added, the package has used a mix of "struct", "Enum".

When also taking nullable reference types into account, the correct generic contraint now seems to be using a combination of "struct" and "Enum".

Fixes #55