Added new "GetNativeLanguageName" extension method for "CultureInfo"
A "CultureInfo" may be based on a language or a language in a specific country. For the latter, the native name will be something like "Dansk (Danmark)", but when based on both a language and a country, the "CultureInfo" will have a reference to parent "CultureInfo" that will only be based on the language, so the native name for this will be "Dansk" instead, which is exactly what we want for this method.

(cherry picked from commit 8577b69319ddf90f970e85644a8928ea0cb82db2)