Fixed misc issues when getting values from JArray
The idea behind these methods is that they should to call, meaning if you GetInt32(0) and the item isn't found, or the value cannot be converted to a 32-bit integer, the method should return a default value instead of throwing an exception. As this wasn't the case, the implementation has been updated to reflect this.

- added check for min and max values
- no longer tries to cast values
- added additional GetType and TryGetType extension methods
- added lots of unit tests
- etc.

Copied from 63c9245ad588a6a140182336edaf4b0441646629