Fixed issue with "EssentialsTime.FromTicks"
Two of the method overloads would internally parse the ticks into a "DateTime" instance, but not specify a "Kind", in which case this falls back to "Unspecified", thereby creating a lot of uncertainty when converting back and forth between time zones. With this commit, the two method overloads will instead use parse the ticks into a "DateTimeOffset" relative to UTC, and then convert to the time zone specified by the user.

Same as 109a493ad7f4c6085aab94db42e7c3a43c49bb9c